### 2021-10-20 #3836 @gkthiruvathukal
@Kevin-Mattheus-Moerman Thanks for the invite.

@fabian-s Beginning with this submission, I am asking all authors to comment on why their proposed package cannot be submitted to the upstream projects already addressing similar issues. In this case, it would be `scikit-learn`, since the "single Lasso" technique is mentioned in your paper as related work.  Please keep in mind that your answer does not preclude consideration in JOSS. It's more about building upon a successful foundation and ensuring sustainability of the work.

Lastly, before I can proceed to review, I will need your help to suggest some prospective reviewers. Given that we are in reduced service mode and some reviewers may  not be available, I need at least FIVE reviewers (without the @ mention) to be listed so I can start inviting folks once I receive your response to the above question(s).
