+++ paper/paper.bib
@@ -1,5 +1,5 @@
 % Encoding: UTF-8
-@Article{Dempster:1972,
+@Article{Dempster1972,
 author = {Dempster, Arthur P.},
 doi = {10.2307/2528966},
 issn = {0006341X},
@@ -10,6 +10,21 @@ publisher = {JSTOR},
 title = {{Covariance selection}},
 volume = {28},
 year = {1972}
+}
+
+@Article{Zhang2021,
+  author   = {Zhang, Ning and Zhang, Yangjing and Sun, Defeng and Toh, Kim-Chuan},
+  journal  = {SIAM J. Math. Data Sci.},
+  title    = {An efficient linearly convergent regularized proximal point algorithm for fused multiple graphical {L}asso problems},
+  year     = {2021},
+  number   = {2},
+  pages    = {524--543},
+  volume   = {3},
+  doi      = {10.1137/20M1344160},
+  fjournal = {SIAM Journal on Mathematics of Data Science},
+  mrclass  = {90C22 (62H22 62J07 90C25)},
+  mrnumber = {4246088},
+  url      = {https://doi-org.eaccess.ub.tum.de/10.1137/20M1344160},
 }
 
 @Article{Kurtz2019,
@@ -39,7 +54,6 @@ keywords = {binary data,convex optimization,gaussian,graphical model,maximum lik
 pages = {485--516},
 title = {{Model selection through sparse maximum likelihood estimation for multivariate gaussian or binary data}},
 url = {http://dl.acm.org/citation.cfm?id=1390696},
-doi = {10.5555/1390681.1390696},
 volume = {9},
 year = {2008}
 }
@@ -230,8 +244,10 @@ year = {2008}
   pages     = {566--577},
   publisher = {Springer International Publishing},
   abstract  = {Network inference is the reverse-engineering problem of inferring graphs from data. With the always increasing availability of data, methods based on probability assumptions that infer multiple intertwined networks have been proposed in literature. These methods, while being extremely flexible, have the major drawback of presenting a high number of hyper-parameters that need to be tuned. The tuning of hyper-parameters, in unsupervised settings, can be performed through criteria based on likelihood or stability. Likelihood-based scores can be easily generalised to the multi hyper-parameters setting, but their computation is feasible only under certain probability assumptions. Differently, stability-based methods are of general application and, on single hyper-parameter, they have been proved to outperform likelihood-based scores. In this work we present a multi-parameters extension to stability-based methods that can be easily applied on complex models. We extensively compared this extension with likelihood-based scores on synthetic Gaussian data. Experiments show that our extension provides a better estimate of models of increasing complexity providing a valuable alternative of existing likelihood-based model selection methods.},
+  doi       = {10.1007/978-3-030-36687-2_47},
   isbn      = {978-3-030-36687-2},
 }
+
 
 @Article{Boyd2011,
   author     = {Boyd, Stephen and Parikh, Neal and Chu, Eric and Peleato, Borja and Eckstein, Jonathan},
