+++ paper/paper.bib
@@ -24,7 +24,6 @@ year = {1972}
   fjournal = {SIAM Journal on Mathematics of Data Science},
   mrclass  = {90C22 (62H22 62J07 90C25)},
   mrnumber = {4246088},
-  url      = {https://doi-org.eaccess.ub.tum.de/10.1137/20M1344160},
 }
 
 @Article{Kurtz2019,
@@ -208,18 +207,6 @@ year = {2008}
   volume  = {12},
 }
 
-@Article{Zhang2019,
-  author      = {Ning Zhang and Yangjing Zhang and Defeng Sun and Kim-Chuan Toh},
-  title       = {An Efficient Linearly Convergent Regularized Proximal Point Algorithm for Fused Multiple Graphical Lasso Problems},
-  year        = {2019},
-  abstract    = {Nowadays, analysing data from different classes or over a temporal grid has attracted a great deal of interest. As a result, various multiple graphical models for learning a collection of graphical models simultaneously have been derived by introducing sparsity in graphs and similarity across multiple graphs. This paper focuses on the fused multiple graphical Lasso model which encourages not only shared pattern of sparsity, but also shared values of edges across different graphs. For solving this model, we develop an efficient regularized proximal point algorithm, where the subproblem in each iteration of the algorithm is solved by a superlinearly convergent semismooth Newton method. To implement the semismooth Newton method, we derive an explicit expression for the generalized Jacobian of the proximal mapping of the fused multiple graphical Lasso regularizer. Unlike those widely used first order methods, our approach has heavily exploited the underlying second order information through the semismooth Newton method. This can not only accelerate the convergence of the algorithm, but also improve its robustness. The efficiency and robustness of our proposed algorithm are demonstrated by comparing with some state-of-the-art methods on both synthetic and real data sets. Supplementary materials for this article are available online.},
-  date        = {2019-02-19},
-  eprint      = {1902.06952v1},
-  eprintclass = {math.OC},
-  eprinttype  = {arXiv},
-  file        = {:http\://arxiv.org/pdf/1902.06952v1:PDF},
-  keywords    = {math.OC},
-}
 
 @Article{Simon2013,
   author    = {Noah Simon and Jerome Friedman and Trevor Hastie and Robert Tibshirani},
