### 2021-10-28 #3865 @whedon
[REVIEW]: GGLasso - a Python package for General Graphical Lasso computation
**Submitting author:** @fabian-sp (<a href="http://orcid.org/0000-0002-0673-9944">Fabian Schaipp</a>)
**Repository:** <a href="https://github.com/fabian-sp/GGLasso" target ="_blank">https://github.com/fabian-sp/GGLasso</a>
**Version:** v0.1.7
