### 2021-10-29 7d020e7
remove redundant computations
