+++ gglasso/solver/ppdna_solver.py
@@ -37,7 +37,7 @@ def get_ppdna_params(ppdna_params = None):
 
 def get_ppa_sub_params_default():
     ppa_sub_params = { 'sigma_t' : 1e3, 
-          'eta' : 1e-5, 'tau' : .2, 'rho' : .5, 'mu' : .1,
+          'eta' : 1e-1, 'tau' : .2, 'rho' : .5, 'mu' : 1e-4,
           'eps_t' : .95, 'delta_t' : .95} 
     
     return ppa_sub_params
@@ -95,13 +95,11 @@ def PPA_subproblem(Omega_t, Theta_t, X_t, S, reg, ppa_sub_params = None, verbose
         W_t = Omega_t - (sigma_t * (S + X_t))  
         V_t = Theta_t + (sigma_t * X_t)
         
-        funY_Xt, gradY_Xt = Y_t( X_t, Omega_t, Theta_t, S, lambda1, lambda2, sigma_t, reg)
+        funY_Xt, Osol, Tsol, (eigD, eigQ) = Y_t( X_t, Omega_t, Theta_t, S, lambda1, lambda2, sigma_t, reg)
+        gradY_Xt = Osol - Tsol
         
-        
-        eigD, eigQ = np.linalg.eigh(W_t)
-        if verbose:
-            print("Eigendecomposition is executed in PPA_subproblem")
-            
+        #eigD, eigQ = np.linalg.eigh(W_t)
+    
         Gamma = construct_gamma(W_t, sigma_t, D = eigD, Q = eigQ)
         W = construct_jacobian_prox_p( (1/sigma_t) * V_t, lambda1 , lambda2, reg)
         
@@ -110,31 +108,34 @@ def PPA_subproblem(Omega_t, Theta_t, X_t, S, reg, ppa_sub_params = None, verbose
         if verbose:
             print("Start CG method")
         
-        D = cg_ppdna(Gamma, eigQ, W, sigma_t, -gradY_Xt, eps = cg_accur, max_iter = 15)
+        D = cg_ppdna(Gamma, eigQ, W, sigma_t, -gradY_Xt, eps = cg_accur, max_iter = 12)
         
         # step 2: line search 
         if verbose:
             print("Start Line search")
         alpha = 1.
-        Y_t_new = Y_t( X_t + alpha * D, Omega_t, Theta_t, S, lambda1, lambda2, sigma_t, reg)[0]
+        Y_t_new, Omega_sol, Theta_sol, _ = Y_t( X_t + alpha * D, Omega_t, Theta_t, S, lambda1, lambda2, sigma_t, reg)
         while Y_t_new < funY_Xt + mu * alpha * Gdot(gradY_Xt , D):
             alpha *= rho
-            Y_t_new = Y_t( X_t + alpha * D, Omega_t, Theta_t, S, lambda1, lambda2, sigma_t, reg)[0]
+            Y_t_new, Omega_sol, Theta_sol, _ = Y_t( X_t + alpha * D, Omega_t, Theta_t, S, lambda1, lambda2,\
+                                                   sigma_t, reg)
             
         # step 3: update variables and check stopping condition
         if verbose:
             print("Update primal-dual variables")
         X_t += alpha * D 
         
-        X_sol = X_t.copy()
+        #### REDUNDANT
+        #X_sol = X_t.copy()
+        # Omega_sol = np.zeros((K,p,p))
+        # eigW, eigV = np.linalg.eigh(Omega_t - sigma_t * (S + X_sol))
+        # for k in np.arange(K):
+        #     _, phip_k, _ = moreau_h(sigma_t , eigW[k,:], eigV[k,:,:])
+        #     Omega_sol[k,:,:] = phip_k
         
[…]
