### 2021-11-02 8b3e3cc
add tikhonov initial
