### 2021-11-03 10a3e7f
make cg faster
