+++ gglasso/solver/ggl_helper.py
@@ -357,7 +357,7 @@ def construct_gamma(A, beta, D = np.array([]), Q = np.array([])):
         
     return Gamma
 
-@njit() 
+@njit()
 def eval_jacobian_phiplus(B, Gamma, Q):
     # numba version of function eval_jacobian_phiplus
     # numba only supports @ for 2D-arrays --> loop through K
@@ -390,7 +390,7 @@ def hessian_Y(D , Gamma, eigQ, W, sigma_t):
     return res
 
 
-def cg_ppdna(Gamma, eigQ, W, sigma_t, b, tol = 1e-6, max_iter = 20, eps = 1e-4):
+def cg_ppdna(Gamma, eigQ, W, sigma_t, b, tol = 1e-6, max_iter = 20, eps = 0.):
     """
     solves the linear system in the PPDNA subproblem
     
@@ -405,26 +405,30 @@ def cg_ppdna(Gamma, eigQ, W, sigma_t, b, tol = 1e-6, max_iter = 20, eps = 1e-4):
     dim = b.shape
     x = np.zeros(dim)
     r = b - hessian_Y(x, Gamma, eigQ, W, sigma_t)
+    r_norm = Gdot(r,r)
     p = r.copy()
+    
     j = 0
+    status = 'max iter'
     
     for j in np.arange(max_iter):
         
-        linp = hessian_Y(p , Gamma, eigQ, W, sigma_t)
+        linp = hessian_Y(p, Gamma, eigQ, W, sigma_t)
+        alpha = r_norm/Gdot(p,linp)
+    
+        x += alpha*p
+        r_norm_old = r_norm
+        r -= alpha*linp
+        r_norm = Gdot(r,r)
         
-        alpha = Gdot(r,r) / Gdot(p, linp)
-        
-        x += alpha * p
-        denom = Gdot(r,r)
-        r -= alpha * linp
-        
-        if np.sqrt(Gdot(r,r)) <= tol:
+        if np.sqrt(r_norm) <= tol:
+            status = f'converged in iter {j}'
             break
         
-        beta = Gdot(r,r)/denom
-        p = r + beta * p 
+        beta = r_norm/r_norm_old
+        p = r + beta*p 
         
-    return x
+    return x, status
 
 def Y_t( X, Omega_t, Theta_t, S, lambda1, lambda2, sigma_t, reg):
     assert np.min(np.array([lambda1, lambda2, sigma_t])) > 0 , "at least one parameter is not positive"
