+++ gglasso/solver/ppdna_solver.py
@@ -25,7 +25,7 @@ def get_ppdna_params(ppdna_params = None):
         assert ppdna_params['max_iter'] > 0
     
     if 'sigma_0' not in p0:
-        ppdna_params['sigma_0'] = 100
+        ppdna_params['sigma_0'] = 100.
     else:
         assert ppdna_params['sigma_0'] > 0
         
@@ -102,11 +102,11 @@ def PPA_subproblem(Omega_t, Theta_t, X_t, S, reg, ppa_sub_params = None, verbose
     
         Gamma = construct_gamma(W_t, sigma_t, D = eigD, Q = eigQ)
         W = construct_jacobian_prox_p( (1/sigma_t) * V_t, lambda1 , lambda2, reg)
-        eps_cg = 1e-3
+        eps_cg = 0.
         # step 1: CG method
         cg_accur = min(eta, np.linalg.norm(gradY_Xt)**(1+tau))
         
-        D = cg_ppdna(Gamma, eigQ, W, sigma_t, -gradY_Xt, tol = cg_accur, max_iter = 12, eps = eps_cg)
+        D, cg_status = cg_ppdna(Gamma, eigQ, W, sigma_t, -gradY_Xt, tol = cg_accur, max_iter = 12, eps = eps_cg)
         
         # step 2: line search 
         alpha = 1.
@@ -211,7 +211,7 @@ def PPDNA(S, lambda1, lambda2, reg, Omega_0, Theta_0 = np.array([]), X_0 = np.ar
     # adds all necessary paramters which are not given as input
     ppdna_params = get_ppdna_params(ppdna_params)
     max_iter = ppdna_params['max_iter']
-    sigma_0 = ppdna_params['sigma_0']
+    sigma_0 = float(ppdna_params['sigma_0'])
     
     ppa_sub_params = get_ppa_sub_params_default()
     ppa_sub_params['sigma_t'] = sigma_0
