### 2021-11-03 65e50ad
reduce computations in cg
