+++ gglasso/solver/ppdna_solver.py
@@ -102,11 +102,10 @@ def PPA_subproblem(Omega_t, Theta_t, X_t, S, reg, ppa_sub_params = None, verbose
     
         Gamma = construct_gamma(W_t, sigma_t, D = eigD, Q = eigQ)
         W = construct_jacobian_prox_p( (1/sigma_t) * V_t, lambda1 , lambda2, reg)
-        eps_cg = 0.
         # step 1: CG method
         cg_accur = min(eta, np.linalg.norm(gradY_Xt)**(1+tau))
         
-        D, cg_status = cg_ppdna(Gamma, eigQ, W, sigma_t, -gradY_Xt, tol = cg_accur, max_iter = 12, eps = eps_cg)
+        D, cg_status = cg_ppdna(Gamma, eigQ, W, sigma_t, -gradY_Xt, tol = cg_accur, max_iter = 10)
         
         # step 2: line search 
         alpha = 1.
@@ -247,15 +246,14 @@ def PPDNA(S, lambda1, lambda2, reg, Omega_0, Theta_0 = np.array([]), X_0 = np.ar
             runtime[iter_t] = end-start
             objective[iter_t] = f(Omega_t, S) + P_val(Omega_t, lambda1, lambda2, reg)
             
-        
+        if verbose:
+            print(out_fmt % (iter_t, eta_P, ppa_sub_params['sigma_t'], sub_info["niter"], sub_info["armijo"], ppa_sub_params['eps_t'], ppa_sub_params['delta_t']))
+         
         if not ppdna_params['sigma_fix']:
             ppa_sub_params['sigma_t'] = 1.3 * ppa_sub_params['sigma_t']
         ppa_sub_params['eps_t'] = 0.95 * ppa_sub_params['eps_t']
         ppa_sub_params['delta_t'] = 0.95 * ppa_sub_params['delta_t']
-            
-        if verbose:
-            print(out_fmt % (iter_t, eta_P, ppa_sub_params['sigma_t'], sub_info["niter"], sub_info["armijo"], ppa_sub_params['eps_t'], ppa_sub_params['delta_t']))
-           
+                    
      
     if eta_P > eps_ppdna:
             status = 'max iterations reached'    
