### 2021-11-03 a49b3c6
replace Gdot by norm
