### 2021-11-03 09f7a1f
fix N bug, add checks in problem
