+++ gglasso/helper/model_selection.py
@@ -199,11 +199,17 @@ def grid_search(solver, S, N, p, reg, l1, l2 = None, w2 = None, method= 'eBIC', 
                 print(SP)
             
             # new best point found
-            if BIC[gamma][g1,g2] < curr_min:
-                print("----------New optimum found in the grid----------")
-                curr_min = BIC[gamma][g1,g2]
-                curr_best = sol.copy()
-    
+            if method == 'eBIC':
+                if BIC[gamma][g1,g2] < curr_min:
+                    print("----------New optimum found in the grid----------")
+                    curr_min = BIC[gamma][g1,g2]
+                    curr_best = sol.copy()
+            elif method == 'AIC':
+                 if AIC[g1,g2] < curr_min:
+                    print("----------New optimum found in the grid----------")
+                    curr_min = AIC[g1,g2]
+                    curr_best = sol.copy()
+        
     # get optimal lambda
     if method == 'AIC':
         AIC[AIC==-np.inf] = np.nan
@@ -220,7 +226,8 @@ def grid_search(solver, S, N, p, reg, l1, l2 = None, w2 = None, method= 'eBIC', 
     
     return stats, ix, curr_best
 
-def K_single_grid(S, lambda_range, N, method = 'eBIC', gamma = 0.3, latent = False, mu_range = None, use_block = True, tol = 1e-7, rtol = 1e-7):
+def K_single_grid(S, lambda_range, N, method = 'eBIC', gamma = 0.3, latent = False, mu_range = None,\
+                  use_block = True, store_all = True, tol = 1e-7, rtol = 1e-7):
     """
     method for doing model selection for K single Graphical Lasso problems, using grid search and AIC/eBIC
     parameters to select: lambda1 (sparsity), mu1 (lowrank, if latent=True)
@@ -248,10 +255,12 @@ def K_single_grid(S, lambda_range, N, method = 'eBIC', gamma = 0.3, latent = Fal
         grid values for mu1. Only needed when latent=True.
     use_block : boolean, optional
         whether to use ADMM on each connected component. Typically, for large and sparse graphs, this is a speedup. Only possible for latent=False.
+    store_all : boolean, optional
+        whether the solution at any grid point is stored. This might be needed if a comparative estimator shall be computed. The default is False.
     tol : float, positive, optional
-            Tolerance for the primal residual used for the solver at each grid point. The default is 1e-7.
+        Tolerance for the primal residual used for the solver at each grid point. The default is 1e-7.
     rtol : float, positive, optional
-            Tolerance for the dual residual used for the solver at each grid point. The default is 1e-7.
+        Tolerance for the dual residual used for the solver at each grid point. The default is 1e-7.
     
 
     Returns
@@ -315,9 +324,11 @@ def K_single_grid(S, lambda_range, N, method = 'eBIC', gamma = 0.3, latent = Fal
             S_k = S[k,:,:].copy()
         
         best, est_k, lr_k, stats_k = single_grid_search(S = S_k, lambda_range = lambda_range, N = N[k], method = method, gamma = gamma, \
-                                                             latent = latent, mu_range = mu_range, use_block = use_block, tol = tol, rtol = rtol)
-        estimates[k] = est_k.copy()
-        lowrank[k] = lr_k.copy()
[…]
