### 2021-11-04 #3865 @fabian-sp
Dear @papachristoumarios and @jameschapman19, we just found that the package `scikit-learn` is missing in `setup.py` but present in `requirements.txt`. Thus if you install `gglasso` by pip, you will need to install `scikit-learn` separately. Then everything works.
Sorry for this, when we tested the installation, we always had an environment with the standard packages and therefore did not spot this. I can fix this right now, but only by creating a new version on Pypi (v0.1.5 then), so would it be possible to change the reviewed version in a later stage?
