### 2021-11-04 4f7bbe2
add assertion for N in SGL
