### 2021-11-04 9cbd72a
replace type by isinstance
