### 2021-11-05 #3865 @fabian-sp
Regarding my above comment, there is version `0.1.5` now on PyPi which fixes the problem with `scikit-learn` as a dependency. I am again sorry for the confusion and hope that it did not affect the review process.
