### 2021-11-05 fabian-sp/GGLasso#27 @papachristoumarios
Warnings when running tests 
I am getting the following warnings when running the tests of the repo (all the tests pass):

```
gglasso/problem.py:19
  /home/mp2242/GGLasso/gglasso/problem.py:19: DeprecationWarning: invalid escape sequence \T
    """

gglasso/problem.py:472
  /home/mp2242/GGLasso/gglasso/problem.py:472: DeprecationWarning: invalid escape sequence \l
    """

gglasso/problem.py:502
  /home/mp2242/GGLasso/gglasso/problem.py:502: DeprecationWarning: invalid escape sequence \l
    """

gglasso/problem.py:690
  /home/mp2242/GGLasso/gglasso/problem.py:690: DeprecationWarning: invalid escape sequence \g
    """

gglasso/solver/admm_solver.py:17
  /home/mp2242/GGLasso/gglasso/solver/admm_solver.py:17: DeprecationWarning: invalid escape sequence \m
    """

gglasso/solver/ext_admm_solver.py:21
  /home/mp2242/GGLasso/gglasso/solver/ext_admm_solver.py:21: DeprecationWarning: invalid escape sequence \m
    """

gglasso/solver/ppdna_solver.py:154
  /home/mp2242/GGLasso/gglasso/solver/ppdna_solver.py:154: DeprecationWarning: invalid escape sequence \m
    """

gglasso/tests/test_solvers.py::test_ADMM_GGL
  /home/mp2242/GGLasso/gglasso/helper/basic_linalg.py:26: NumbaPerformanceWarning: '@' is faster on contiguous arrays, called on (array(float64, 2d, A), array(float64, 2d, A))
    res += Sdot(X[k,:,:], Y[k,:,:])

gglasso/tests/test_solvers.py::test_admm_ppdna_ggl
  /home/mp2242/GGLasso/gglasso/solver/ggl_helper.py:262: NumbaPerformanceWarning: '@' is faster on contiguous arrays, called on (array(float64, 2d, A), array(float64, 2d, F))
    ij_entry = jacobian_prox_phi(X[:,i,j] , l1 , l2, reg)

gglasso/tests/test_solvers.py::test_admm_ppdna_ggl
  /home/mp2242/GGLasso/gglasso/solver/ggl_helper.py:262: NumbaPerformanceWarning: '@' is faster on contiguous arrays, called on (array(float64, 2d, C), array(float64, 2d, A))
    ij_entry = jacobian_prox_phi(X[:,i,j] , l1 , l2, reg)

gglasso/tests/test_solvers.py::test_admm_ppdna_ggl
  /home/mp2242/GGLasso/gglasso/solver/ggl_helper.py:417: NumbaPerformanceWarning: '@' is faster on contiguous arrays, called on (array(float64, 2d, A), array(float64, 1d, A))
    r = b - hessian_Y(x, Gamma, eigQ, W, sigma_t)

gglasso/tests/test_solvers.py::test_admm_ppdna_ggl
  /home/mp2242/GGLasso/gglasso/solver/ggl_helper.py:417: NumbaPerformanceWarning: '@' is faster on contiguous arrays, called on (array(float64, 2d, F), array(float64, 2d, A))
    r = b - hessian_Y(x, Gamma, eigQ, W, sigma_t)
```

I think that the warnings can be easily fixed.

Thank you in advance
