### 2021-11-05 fabian-sp/GGLasso#28 @papachristoumarios
Compatibility with sklearn API
I have noticed that in the project's source code and documentation methods that are present in similar sklearn classes such as [GraphicalLasso](https://scikit-learn.org/stable/modules/generated/sklearn.covariance.GraphicalLasso.html#sklearn.covariance.GraphicalLasso) have methods (e.g. the `fit` method) which can be found in the current project but under different names and specifications.

For example, [here](https://github.com/fabian-sp/GGLasso/blob/master/paper/paper.md#installation-and-problem-initiation) the GGL class can be created separately and then the data can be provided via a `fit` method (like they do in sklearn) instead of defining a `gglasso_problem` with the data and then calling the `solve` method without any arguments.

Thus, I think that the software can be improved by having method naming and functionality that is similar to sklearn (since sklearn is widely used by the data science / ML community).
