### 2021-11-05 1934b2f
initial thresholding functions
