### 2021-11-05 #3865 @papachristoumarios
@gkthiruvathukal I have finished my review. In general the paper presents a complete body of work, is very well written and has thorough evaluation. I went through all the criteria and I couldn't find something that was missing. I have also opened two issues in the repo regarding the improvement of the software:

* https://github.com/fabian-sp/GGLasso/issues/27
* https://github.com/fabian-sp/GGLasso/issues/28

The former can be addressed easily. The second is more design-oriented and I think will improve the usability of the software and code portability.

Thank you
