### 2021-11-08 fabian-sp/GGLasso#27 @fabian-sp
Thanks for raising the issue.
1) I did not get the `DeprecationWarning` but I guess they can be fixed with an additional `\`?
2) Regarding the `NumbaPerformanceWarning`: i tried to fix these, but got the warning even when explicitly converting into an contiguous array before calling the function (which would be anyways a rather messy solution). I concluded from https://stackoverflow.com/questions/54026429/numba-raises-performance-warning-when-using-eventhough-array-are-contiguous that this could be a numba internal issue. Do you know a solution for this from prior experience?
