### 2021-11-08 fabian-sp/GGLasso#27 @papachristoumarios
Hi @fabian-sp,

thank you for the reply. See my responses below:

> 1. I did not get the `DeprecationWarning` but I guess they can be fixed with an additional `\`?

Yes! I think it can be fixed this way.

> 2. Regarding the `NumbaPerformanceWarning`: i tried to fix these, but got the warning even when explicitly converting into an contiguous array before calling the function (which would be anyways a rather messy solution). I concluded from https://stackoverflow.com/questions/54026429/numba-raises-performance-warning-when-using-eventhough-array-are-contiguous that this could be a numba internal issue. Do you know a solution for this from prior experience?

Since the solution becomes messy and the warning persists after you have changed, I do not see a point in changing the initial code in this occasion (we can also ask the second reviewer for that). Moreover, I do not know any solution from prior experience.

Thank you
