### 2021-11-10 fabian-sp/GGLasso#28 @fabian-sp
Thank you for the input! Indeed we had discussed compatibility with sklearn when designing the package. One of the core ideas of GGLasso was to have one entry point (i.e. `glasso_problem`) for all problem formulations. We derive the problem formulation from the shape of the input data `S` (and additional arguments such as `latent` and `reg`) which is why we need the input already when creating the problem instance right now. Moreover, as regularization parameters are part of the problem and depend on the input (e.g. via scaling), we thought that it also makes sense to have the input data as part of the problem (even though this differs from sklearn).
Regarding the second point, do in understand you correctly that you propose to rename the `.solve` method to `.fit`? This would be an easy change (even though I personally think that problems are solved rather than fitted).
