### 2021-11-11 19929d8
remove redundant comp
