### 2021-11-11 6f75d54
add thresholding to single grid
