### 2021-11-11 fabian-sp/GGLasso#27 @fabian-sp
Hi @papachristoumarios ,

the issue is adressed in https://github.com/fabian-sp/GGLasso/pull/29.

Thank you for your input.
