### 2021-11-16 fabian-sp/GGLasso#28 @jameschapman19
I think I'm inclined to agree with @papachristoumarios that this would be a nice change and could even buy you some functionality for free from scikit-learn.

```
class FabianGraphicalLasso(EmpiricalCovariance):
    def __init__(self, score_type='AIC'):
        super().__init__()
        self.score_type = score_type

    def fit(self, X: Union[np.ndarray, Iterable[np.ndarray]], y=None):
        self.N = X.shape[0]
        self.S = empirical_covariance(X, assume_centered=self.assume_centered)
        self._derive_problem_formulation()
        self._solve()  # sklearn has conventions for fitted parameters beginning ._fitted_parameter so this would set these
        # so that
        return self

    def score(self, X_test, y=None):
        S_test = empirical_covariance(X_test - self.location_, assume_centered=True)
        if self.score_type == 'AIC':
            res = self._aic(S_test)
        else:
            res = self._ebic(S_test)
        return res

    def _derive_problem_formulation(self):
        pass

    def _solve(self):
        pass

    def _aic(self, S):
        pass

    def _ebic(self, S):
        pass
```
