### 2021-11-16 fabian-sp/GGLasso#28 @jameschapman19
I should add (having written a package for multiview data myself) that there is one major frustration with scikit-learn api which is that it expects X,y to both be arrays rather than iterables of arrays which is what we would like!

Worth checking out mvlearn if you aren't familiar. They seem to have a few of the scikit developers.
