### 2021-11-16 #3865 @jameschapman19
I'm not sure how to edit the checklist but I have checked them all off.

Firstly cool package and it's nice to be familiar enough with the area to be able to review this so thankyou for the opportunity.

A couple of bits from me:

- I wonder if the tests under test_problem could have some sort of basic assertion (this is a problem I have had with my own recent JOSS submission!) I assume if the problems are seeded you can assert the solutions as they seem low dimensional.
- Agree with @papachristoumarios about the scikit API. This is to say that the package in current form is sensibly designed and I get the logic of the problem object but... people love a familiar API and probably long run it helps grab users. I also think you might be able to get some free code logic (pipeline stuff for CV) this way too - at least for the SGL. It's not necessary for me to recommend the software though.
