### 2021-11-16 #3865 @jameschapman19
One more:

- Could solve have a verbosity=False option
