### 2021-11-16 fabian-sp/GGLasso#28 @jameschapman19
Actually thinking about it X could be your data and y could be labels k for the multiple case!
