### 2021-11-16 fabian-sp/GGLasso#28 @jameschapman19
So the really cool thing about this route is you could substitute most of your gridsearch functionality for the scikit-learn functionality (i.e. could use random search, gridsearch)
