### 2021-11-16 5b2673a
add verbosity and warnings
