### 2021-11-16 ff8b8ee
add assertion in test problems
