+++ gglasso/tests/test_problem.py
@@ -2,8 +2,9 @@
 author: Fabian Schaipp
 """
 import numpy as np
+from numpy.testing import assert_array_almost_equal
 
-from gglasso.helper.data_generation import time_varying_power_network, group_power_network,sample_covariance_matrix
+from gglasso.helper.data_generation import time_varying_power_network, group_power_network, sample_covariance_matrix, generate_precision_matrix
 from gglasso.problem import glasso_problem
 from gglasso.helper.ext_admm_helper import construct_trivial_G
 
@@ -51,30 +52,30 @@ def template_problem_MGL(S, N, reg = 'GGL', latent = False, G = None):
     _ = P.solution.calc_ebic(gamma = 0.1)
     P.solution.calc_adjacency()
     
-    return
+    return P
 
 def test_GGL():
     Sigma, Theta = group_power_network(p, K, M)    
     S, samples = sample_covariance_matrix(Sigma, N)
-    template_problem_MGL(S, N, reg = 'GGL', latent = False)   
+    _ = template_problem_MGL(S, N, reg = 'GGL', latent = False)   
     return
 
 def test_GGL_latent():
     Sigma, Theta = group_power_network(p, K, M)    
     S, samples = sample_covariance_matrix(Sigma, N)
-    template_problem_MGL(S, N, reg = 'GGL', latent = True)
+    _ = template_problem_MGL(S, N, reg = 'GGL', latent = True)
     return
 
 def test_FGL():
     Sigma, Theta = time_varying_power_network(p, K, M)
     S, samples = sample_covariance_matrix(Sigma, N)
-    template_problem_MGL(S, N, reg = 'FGL', latent = False)
+    _ = template_problem_MGL(S, N, reg = 'FGL', latent = False)
     return
 
 def test_FGL_latent():
     Sigma, Theta = time_varying_power_network(p, K, M)
     S, samples = sample_covariance_matrix(Sigma, N)
-    template_problem_MGL(S, N, reg = 'FGL', latent = True)
+    _ = template_problem_MGL(S, N, reg = 'FGL', latent = True)
     return
 
 def test_GGL_ext():
@@ -86,7 +87,7 @@ def test_GGL_ext():
         Sdict[k] = S[k,:,:].copy()
         
     G = construct_trivial_G(p, K)
-    template_problem_MGL(Sdict, N, reg = 'GGL', latent = False, G = G)
+    _ = template_problem_MGL(Sdict, N, reg = 'GGL', latent = False, G = G)
     return
 
 def test_GGL_ext_latent():
@@ -98,7 +99,7 @@ def test_GGL_ext_latent():
         Sdict[k] = S[k,:,:].copy()
         
     G = construct_trivial_G(p, K)
-    template_problem_MGL(Sdict, N, reg = 'GGL', latent = True, G = G)
+    _ = template_problem_MGL(Sdict, N, reg = 'GGL', latent = True, G = G)
     return
 
 ###############################################################
@@ -137,46 +138,55 @@ def template_problem_SGL(S, N, latent = False):
     #tmp = P.modelselect_stats.copy()
     _ = P.solution.calc_ebic(gamma = 0.1)
     P.solution.calc_adjacency()
-    return
+    return P
 
 def test_SGL():
-    Sigma, Theta = group_power_network(p, K = 1, M = 2)
-    S, samples = sample_covariance_matrix(Sigma, N); S = S[0,:,:]  
-    template_problem_SGL(S, N, latent = False)
+    Sigma, Theta = group_power_network(p, K = 1, M = 2, seed = 1234)
[…]
