### 2021-11-16 #3865 @fabian-sp
@jameschapman19  Thanks for your input! I adressed the points not related to `sklearn` in this PR: https://github.com/fabian-sp/GGLasso/pull/29
