### 2021-11-16 #3865 @papachristoumarios
@gkthiruvathukal The author has addressed the issues we referred to within the submission. The issue regarding the API compatibility with sklearn is in my opinion more of a design aesthetic issue (which can have some functionality benefits as @jameschapman19 and increase visibility) but for me, it's also not necessary for recommending acceptance of the software.
