### 2021-11-17 de977fa
remove duplicate bib
