+++ paper/paper.bib
@@ -119,14 +119,6 @@ year = {2008}
   publisher = {Public Library of Science ({PLoS})},
 }
 
-@Article{Kurtz2019,
-  author    = {Zachary D. Kurtz and Richard Bonneau and Christian L. Müller},
-  title     = {Disentangling microbial associations from hidden environmental and technical factors via latent graphical models},
-  year      = {2019},
-  month     = {dec},
-  doi       = {10.1101/2019.12.21.885889},
-  publisher = {Cold Spring Harbor Laboratory},
-}
 
 @Article{Ma2013,
   author    = {Shiqian Ma and Lingzhou Xue and Hui Zou},
