### 2021-11-18 c85c57d
adding a standart version attribute to gglasso
