+++ setup.py
@@ -1,6 +1,5 @@
 import io
 import os
-import re
 import gglasso
 from setuptools import setup, find_packages
 
