### 2021-11-22 #3865 @fabian-sp
@gkthiruvathukal The Github release tag is `v0.1.7`.
The Zenodo archive ihas the DOI `10.5281/zenodo.5718440` (link: https://zenodo.org/record/5718440#.YZuuRJHMI5k)

Let me know if the metadata is correct.
