### 2021-11-23 5ed0661
fix transpose in Sdot
