### 2021-11-26 120c777
add prox of functional glasso
