+++ gglasso/solver/functional_sgl_admm.py
@@ -1,0 +1,237 @@
+"""
+@author: Fabian Schaipp
+"""
+
+import numpy as np
+import time
+import warnings
+
+from .ggl_helper import prox_od_1norm, phiplus, prox_rank_norm
+
+
+def ADMM_FSGL(S, lambda1, M, Omega_0, Theta_0=np.array([]), X_0=np.array([]),
+             rho=1., max_iter=1000, tol=1e-7, rtol=1e-4,\
+             update_rho=True, verbose=False, measure=False, latent=False, mu1=None):
+    """
+    This is an ADMM solver for the (Latent variable) Functional Single Graphical Lasso problem (FSGL).
+    If ``latent=False``, this function solves
+
+    .. math::
+       \min_{\Omega, \Theta \in \mathbb{S}^p_{++}} - \log \det \Omega + \mathrm{Tr}(S\Omega) + \lambda \|\Theta\|_{1,od}
+
+       s.t. \quad \Omega = \Theta
+
+    If ``latent=True``, this function solves
+
+    .. math::
+       \min_{\Omega, \Theta, L \in \mathbb{S}^p_{++}} - \log \det (\Omega) + \mathrm{Tr}(S \Omega) + \lambda_1 \|\Theta\|_{1,od} + \mu_1 \|L\|_{\star}
+
+       s.t. \quad \Omega = \Theta - L
+
+    Note:
+        * Typically, ``sol['Omega']`` is positive definite and ``sol['Theta']`` is sparse.
+        * We use scaled ADMM, i.e. X are the scaled (with 1/rho) dual variables for the equality constraint.
+    Parameters
+    ----------
+    S : array (p,p)
+        empirical covariance matrix. Needs to be symmetric and positive semidefinite.
+    lambda1 : float, positive
+        sparsity regularization parameter.
+    M : int
+        each variable is a M-dimensional representation.
+    Omega_0 : array (p,p)
+        starting point for the Omega variable. Choose ``np.eye(p)`` if no better starting point is known.
+    Theta_0 : array (p,p), optional
+        starting point for the Theta variable. If not specified, it is set to the same as Omega_0.
+    X_0 : array (p,p), optional
+        starting point for the X variable. If not specified, it is set to zero array.
+    rho : float, positive, optional
+        step size paramater for the augmented Lagrangian in ADMM. The default is 1. Tune this parameter for optimal performance.
+    max_iter : int, optional
+        maximum number of iterations. The default is 1000.
+    tol : float, positive, optional
+        tolerance for the primal residual. See "Distributed Optimization and Statistical Learning via the Alternating Direction Method of Multipliers", Boyd et al. for details.
+        The default is 1e-7.
+    rtol : float, positive, optional
+        tolerance for the dual residual. The default is 1e-4.
+    update_rho : boolean, optional
+        Whether the penalty parameter ``rho`` is updated, see Boyd et al. page 20-21 for details. The default is True.
+    verbose : boolean, optional
+        verbosity of the solver. The default is False.
+    measure : boolean, optional
+        turn on/off measurements of runtime per iteration. The default is False.
+    latent : boolean, optional
+        Solve the SGL with or without latent variables (see above for the exact formulations).
+        The default is False.
+    mu1 : float, positive, optional
+        low-rank regularization parameter. Only needs to be specified if latent=True.
+    Returns
+    -------
+    sol : dict
[…]
