+++ gglasso/problem.py
@@ -569,7 +569,7 @@ class glasso_problem:
             # LATENT VARIABLES --> FIRST STAGE lambda1/mu1 for each instance
             ############################### 
             if self.latent:
-                est_uniform, est_indv, stage1_statistics = K_single_grid(S = self.S, lambda_range = self.modelselect_params['lambda1_range'], N = self.N, method = method,\
+                _, self.est_indv, stage1_statistics = K_single_grid(S = self.S, lambda_range = self.modelselect_params['lambda1_range'], N = self.N, method = method,\
                                                                   gamma = gamma, latent = self.latent, mu_range = self.modelselect_params['mu1_range'],
                                                                   use_block = True, store_all = False, tol = tol, rtol= rtol)            
                 
