### 2021-11-27 90c8069
revert structure of estimators
