### 2021-11-28 6b1cb1b
finish solver and add tests
