+++ gglasso/solver/functional_sgl_admm.py
@@ -6,7 +6,7 @@ import numpy as np
 import time
 import warnings
 
-from .ggl_helper import prox_od_1norm, phiplus, prox_rank_norm
+from .ggl_helper import prox_sum_Frob, phiplus, prox_rank_norm
 
 
 def ADMM_FSGL(S, lambda1, M, Omega_0, Theta_0=np.array([]), X_0=np.array([]),
@@ -14,36 +14,39 @@ def ADMM_FSGL(S, lambda1, M, Omega_0, Theta_0=np.array([]), X_0=np.array([]),
              update_rho=True, verbose=False, measure=False, latent=False, mu1=None):
     """
     This is an ADMM solver for the (Latent variable) Functional Single Graphical Lasso problem (FSGL).
+    It solves a SGL problem for the case when each of the ``p`` variables has an ``M``-dimensional functional representation (e.g. Fourier transform).
+    Therefore, the input data has the shape ``(p*M,p*M)``.  
+    
     If ``latent=False``, this function solves
 
     .. math::
-       \min_{\Omega, \Theta \in \mathbb{S}^p_{++}} - \log \det \Omega + \mathrm{Tr}(S\Omega) + \lambda \|\Theta\|_{1,od}
+       \min_{\Omega, \Theta \in \mathbb{S}^{p\cdot M}_{++}} - \log \det \Omega + \mathrm{Tr}(S\Omega) + \lambda_1 \sum_{j\neq l} \|\Theta_{jl}^M\|_{F}
 
        s.t. \quad \Omega = \Theta
 
     If ``latent=True``, this function solves
 
     .. math::
-       \min_{\Omega, \Theta, L \in \mathbb{S}^p_{++}} - \log \det (\Omega) + \mathrm{Tr}(S \Omega) + \lambda_1 \|\Theta\|_{1,od} + \mu_1 \|L\|_{\star}
+       \min_{\Omega, \Theta, L \in \mathbb{S}^{p\cdot M}_{++}} - \log \det (\Omega) + \mathrm{Tr}(S \Omega) + \lambda_1 \sum_{j\neq l} \|\Theta_{jl}^M\|_{F} + \mu_1 \|L\|_{\star}
 
        s.t. \quad \Omega = \Theta - L
 
     Note:
         * Typically, ``sol['Omega']`` is positive definite and ``sol['Theta']`` is sparse.
-        * We use scaled ADMM, i.e. X are the scaled (with 1/rho) dual variables for the equality constraint.
+        * We use scaled ADMM, i.e. X are the scaled (with ``1/rho``) dual variables for the equality constraint.
     Parameters
     ----------
-    S : array (p,p)
+    S : array (p*M,p*M)
         empirical covariance matrix. Needs to be symmetric and positive semidefinite.
     lambda1 : float, positive
-        sparsity regularization parameter.
+        regularization parameter for the Frobenius norm of the MxM subblocks.
     M : int
-        each variable is a M-dimensional representation.
-    Omega_0 : array (p,p)
-        starting point for the Omega variable. Choose ``np.eye(p)`` if no better starting point is known.
-    Theta_0 : array (p,p), optional
+        Dimension of the functional representation. See "Functional Graphical Models", Qiao et al. for details.
+    Omega_0 : array (p*M,p*M)
+        starting point for the Omega variable. Choose ``np.eye(p*M)`` if no better starting point is known.
+    Theta_0 : array (p*M,p*M), optional
[…]
