### 2021-11-29 2e51cc2
add multiple thresholding
