+++ gglasso/helper/model_selection.py
@@ -12,18 +12,20 @@ from .ext_admm_helper import get_K_identity as id_dict
 from ..solver.single_admm_solver import ADMM_SGL, block_SGL
 
 
+# default parameters
+DEFAULT_GAMMAS = [0.1, 0.3, 0.5, 0.7]
+TAU_MIN = 1e-12
+N_TAU = 20
       
 def lambda_parametrizer(l1 = 0.05, w2 = 0.5):
     """transforms given l1 and w2 into the respective l2"""
     a = 1/np.sqrt(2)
     l2 = (w2*l1)/(a*(1-w2))
-
     return l2
 
 def map_l_to_w(l1, l2):
     w1 = l1 + (1/np.sqrt(2)) *l2
-    w2 = l2/(w1*np.sqrt(2))
-    
+    w2 = l2/(w1*np.sqrt(2))   
     return (w1,w2)
     
 def lambda_grid(l1, l2 = None, w2 = None):
@@ -121,7 +123,7 @@ def grid_search(solver, S, N, p, reg, l1, l2 = None, w2 = None, method= 'eBIC', 
     AIC = np.zeros((grid1, grid2))
     AIC[:] = np.nan
     
-    gammas = [0.1, 0.3, 0.5, 0.7]
+    gammas = DEFAULT_GAMMAS
     gammas.append(gamma)
     gammas = list(set(gammas))
      
@@ -181,7 +183,9 @@ def grid_search(solver, S, N, p, reg, l1, l2 = None, w2 = None, method= 'eBIC', 
                          
             # warm start
             kwargs['Omega_0'] = Omega_sol
-                            
+                  
+            #########################################
+            
             # store diagnostics
             AIC[g1,g2] = aic(S, Theta_sol, N)
             for g in gammas:
@@ -210,8 +214,7 @@ def grid_search(solver, S, N, p, reg, l1, l2 = None, w2 = None, method= 'eBIC', 
     # get optimal lambda
     if method == 'AIC':
         AIC[AIC==-np.inf] = np.nan
-        ix= np.unravel_index(np.nanargmin(AIC), AIC.shape)
-    
+        ix= np.unravel_index(np.nanargmin(AIC), AIC.shape)   
     elif method == 'eBIC':    
         for g in gammas:
             BIC[g][BIC[g]==-np.inf] = np.nan
@@ -224,7 +227,7 @@ def grid_search(solver, S, N, p, reg, l1, l2 = None, w2 = None, method= 'eBIC', 
     return stats, ix, curr_best
 
 def K_single_grid(S, lambda_range, N, method = 'eBIC', gamma = 0.3, latent = False, mu_range = None,\
-                  use_block = True, store_all = True, tol = 1e-7, rtol = 1e-7):
+                  thresholding = False, use_block = True, store_all = True, tol = 1e-7, rtol = 1e-7):
     """
     method for doing model selection for K single Graphical Lasso problems, using grid search and AIC/eBIC
     parameters to select: lambda1 (sparsity), mu1 (lowrank, if latent=True)
@@ -251,6 +254,8 @@ def K_single_grid(S, lambda_range, N, method = 'eBIC', gamma = 0.3, latent = Fal
         whether to model latent variables or not. The default is False.
     mu_range : array, optional
         grid values for mu1. Only needed when latent=True.
+    thresholding : boolean, optional
+        whether to tune a thresholded estimator for each (lambda1,mu1) pair. See https://arxiv.org/pdf/2104.06389v1.pdf for details.
     use_block : boolean, optional
         whether to use ADMM on each connected component. Typically, for large and sparse graphs, this is a speedup. Only possible for latent=False.
[…]
