### 2021-11-29 77787ed
remove redundancies
