### 2021-12-10 #3865 @danielskatz
@fabian-sp - I've suggested a bunch of changes in https://github.com/fabian-sp/GGLasso/pull/32
Notable, if you say in the text `following [@ref]`, this should be just `following @ref` with `[@ref]` just used when the reference isn't explicit, e.g. `using the X method [@refX]`  Please merge the changes, ask whedon to regenerate the pdf (`@whedon generate pdf`) then carefully check the pdf to make sure the refs are ok, and let me know when you are done, then we can proceed to final publishing
